requirements.txt
# Statistical analysis project - install with: pip install -r requirements.txt
# For PDF report: pip install fpdf2 then run python generate_report_pdf.py
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scipy>=1.7.0
jupyter>=1.0.0
notebook>=6.4.0
kagglehub>=0.2.0
fpdf2>=2.7.0
